last

+++ 0x00-hello_world/6-size.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
-
+#include <unistd.h>
 /**
  * main - Entry point
  *
@@ -14,10 +14,10 @@ int main(void)
 	long long int b;
 	float f;
 
-	printf("Size of a char : %lu byte(s)\n", (unsigned char)sizeof(c);
-	printf("Size of an int: %lu byte(s)\n", (unsigned int)sizeof(i);
+	printf("Size of a char : %lu byte(s)\n", (unsigned char)sizeof(c));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(i));
 	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(a));
 	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(b));
-	printf("Size of a float: %lu byte(s)\n", (unsigned float)sizeof(f);
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
